Restructure application file structure
