+++ lib/msal-electron-proof-of-concept/src/ServerRequest/Error/AuthorizationCodeRequestError.ts
@@ -1,7 +1,7 @@
 // Copyright (c) Microsoft Corporation. All rights reserved.
 // Licensed under the MIT License
 
-import { AuthError } from './AuthError';
+import { AuthError } from '../../AppConfig/Error/AuthError';
 
 /**
  * The AuthorizationCodeRequestErrorMessage object holds all
