+++ lib/msal-electron-proof-of-concept/src/ServerRequest/ServerRequestParameters.ts
@@ -1,7 +1,7 @@
 // Copyright (c) Microsoft Corporation. All rights reserved.
 // Licensed under the MIT License.
 
-import { Authority } from '../Authority/Authority';
+import { Authority } from '../AppConfig/Authority/Authority';
 
 /**
  * ServerRequestParamters is the base class from which AuthorizationCodeRequestParameters
