+++ lib/msal-electron-proof-of-concept/src/ServerRequest/TokenRequestOptions.ts
