+++ lib/msal-electron-proof-of-concept/src/ServerResponse/QueryParameters.ts
